requirements.txt
django >= 3.2, < 3.3
django-cors-headers==3.13.0
djangorestframework >= 3.12, < 3.13
drf-yasg==1.21.3
swagger-spec-validator==2.7.4
user-agents==2.2.0